Correctly memoize subcalculations
Updates the state with the cache at the time of saving, rather than
using the cache retrieved before recursing, as doing the latter means
the cache does not contain any elements calculated during the recursion.

Note that correctly using the cache does not appear to actually speed up
the calculation, presumably because sufficiently few calculations are
re-used to make up for the time spent updating the cache.