Remove assumption that all inputs are prime
Day 13 part 2 takes advantage of the fact that, if a number n matches
constraints for two numbers a and b, all numbers matching those
constraints will be of the form n + k * lcm(a, b) for integral k.
However, the previous version didn't actually calculate the lcm; it
relied on the fact that the actual input has all prime numbers, and so
the lcm is just the product. This commit actually calculates the lcm,
as the input values all being prime is not guaranteed by the problem
description (happily, this does not seem to make the solution any
slower).